sweep: remove non-wallet keys from sweep UTXOs
This was made inconsistent when singlesig sweeping was added, and then
broken when client side sweeping was made the default for multisig.

e.g. ga_session::keys_from_utxo currently relies on this. Although that
will change in the future, revert the retained data format to match the
previous expectations.