taproot: Ensure required prevout info is available when signing p2tr
Not all paths leading to tx signing provide us with input scriptpubkeys,
and some (such as liquidex) don't provide us with e.g. prevout asset tags.

Augment the tx inputs in these cases with the required signing data so
that it is available for both internal and external signers.

Remove the prevous solution where the signer was responsible for
fetching this data, since it a) requires all signers including external
ones to be changed and b) is inefficient when the prevouts have already
been fetched before signing in many cases.