network: use staging pin server for localtest and testnet/signet envs
Currently test setups are spamming the production pinserver which is
obviously not ideal.

This may be slightly disruptive to test accounts (will need to re-setup
their pin data), but has to be fixed at some point.

Note tor is not yet available but will be added back when it is.